* include/cygwin/types.h: Fix erroneous definition of ino_t from 	2003-04-28.

+++ winsup/cygwin/include/cygwin/types.h
@@ -90,7 +90,7 @@ typedef __gid16_t gid_t;
 
 #ifndef __ino_t_defined
 #define __ino_t_defined
-typedef unsigned long long ino_t;
+typedef unsigned long ino_t;
 #endif /*__ino_t_defined*/
 
 #ifndef __BIT_TYPES_DEFINED
